DPDK 18.02, X520-DA2 82599 NIC
changes in rte_eth_dev initialization in pmd.c, for working with
X520-DA2 ntuple and fdir filter

+++ framework/src/common/errors.rs
@@ -12,7 +12,7 @@ error_chain! {
             description("Failed to initialize port")
             display("Failed to initialize port: {}", port)
         }
-        FailedToInitializeKni(port_id: u8) {
+        FailedToInitializeKni(port_id: u16) {
             description("Failed to initialize kni i/f")
             display("Failed to initialize kni i/f: port_id={}", port_id)
         }
